backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
elasticsearch==8.11.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
bcrypt==4.1.2
requests==2.31.0
pandas==2.1.3
snowflake-connector-python==3.6.0
prometheus-client==0.19.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==20.1.0
